tests: rename test files to match type being tested
