+++ Testing/CarbonateTests/TwoWay/RespondSubscriptionTests.cs
@@ -1,18 +1,18 @@
-﻿// <copyright file="RespondReactorTests.cs" company="KinsonDigital">
+﻿// <copyright file="RespondSubscriptionTests.cs" company="KinsonDigital">
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace CarbonateTests.OneWay;
+namespace CarbonateTests.TwoWay;
 
-using Carbonate.OneWay;
+using Carbonate.TwoWay;
 using FluentAssertions;
 using Moq;
 using Xunit;
 
 /// <summary>
-/// Tests the <see cref="RespondSubscription{TOut}"/> class.
+/// Tests the <see cref="RespondSubscription{TIn,TOut}"/> class.
 /// </summary>
-public class RespondReactorTests
+public class RespondSubscriptionTests
 {
     #region Constructor Tests
     [Fact]
@@ -22,7 +22,7 @@ public class RespondReactorTests
         var id = Guid.NewGuid();
 
         // Act
-        var sut = new RespondSubscription<string>(id);
+        var sut = new RespondSubscription<int, string>(id);
 
         // Assert
         sut.Id.Should().Be(id);
@@ -35,7 +35,7 @@ public class RespondReactorTests
         var id = Guid.NewGuid();
 
         // Act
-        var sut = new RespondSubscription<string>(id, "test-name");
+        var sut = new RespondSubscription<int, string>(id, "test-name");
 
         // Assert
         sut.Name.Should().Be("test-name");
@@ -44,47 +44,62 @@ public class RespondReactorTests
 
     #region Method Tests
     [Fact]
-    public void OnRespond_WhenUnsubscribed_DoesNotInvokeOnRespondAction()
+    public void OnRespond_WhenUnsubscribed_ReturnsCorrectDefaultResult()
     {
         // Arrange
-        var totalActionInvokes = 0;
-        var sut = new RespondSubscription<string>(
-            It.IsAny<Guid>(),
-            It.IsAny<string>(),
-            onRespond: () =>
-            {
-                totalActionInvokes++;
-                return It.IsAny<string>();
-            });
-
+        var sut = new RespondSubscription<int, int>(Guid.NewGuid(),
+            onRespond: _ => 456);
         sut.OnUnsubscribe();
 
         // Act
-        _ = sut.OnRespond();
+        var actual = sut.OnRespond(123);
 
         // Assert
-        totalActionInvokes.Should().Be(0);
+        actual.Should().Be(0);
     }
 
     [Fact]
-    public void OnRespond_WhenNotUnsubscribed_InvokesOnRespondAction()
+    public void OnRespond_WhenDataIsNull_ThrowsException()
     {
         // Arrange
-        var totalActionInvokes = 0;
-        var sut = new RespondSubscription<string>(
-            It.IsAny<Guid>(),
-            It.IsAny<string>(),
-            onRespond: () =>
-            {
-                totalActionInvokes++;
-                return "return-value";
-            });
+        var sut = new RespondSubscription<object, int>(Guid.NewGuid());
 
         // Act
-        _ = sut.OnRespond();
+        var act = () => sut.OnRespond(null);
 
         // Assert
-        totalActionInvokes.Should().Be(1);
+        act.Should().Throw<ArgumentNullException>()
+            .WithMessage("The parameter must not be null. (Parameter 'data')");
+    }
+
+    [Fact]
+    public void OnRespond_WhenOnRespondDataIsNull_ReturnsCorrectDefaultResult()
+    {
+        // Arrange
+        var sut = new RespondSubscription<int, object>(Guid.NewGuid(),
+            onRespond: _ => null);
+
+        // Act
+        var actual = sut.OnRespond(123);
+
+        // Assert
+        actual.Should().BeNull();
+    }
+
+    [Fact]
+    public void OnRespond_WhenOnRespondDataIsNotNull_ReturnsCorrectResult()
+    {
+        // Arrange
+        var obj = new object();
+        var sut = new RespondSubscription<int, object>(Guid.NewGuid(),
+            onRespond: _ => obj);
+
+        // Act
+        var actual = sut.OnRespond(123);
+
+        // Assert
+        actual.Should().NotBeNull();
+        actual.Should().BeSameAs(obj);
     }
 
     [Fact]
@@ -93,7 +108,7 @@ public class RespondReactorTests
         // Arrange
         var totalActionInvokes = 0;
 
-        var sut = new RespondSubscription<string>(
+        var sut = new RespondSubscription<int, string>(
             It.IsAny<Guid>(),
             It.IsAny<string>(),
             onUnsubscribe: () => totalActionInvokes++);
@@ -113,7 +128,7 @@ public class RespondReactorTests
         // Arrange
         var totalActionInvokes = 0;
 
-        var sut = new RespondSubscription<string>(
+        var sut = new RespondSubscription<int, string>(
             It.IsAny<Guid>(),
             It.IsAny<string>(),
             onError: _ => totalActionInvokes++);
@@ -133,7 +148,7 @@ public class RespondReactorTests
         // Arrange
         var totalActionInvokes = 0;
 
-        var sut = new RespondSubscription<string>(
+        var sut = new RespondSubscription<int, string>(
             It.IsAny<Guid>(),
             It.IsAny<string>(),
             onError: _ => totalActionInvokes++);
@@ -153,7 +168,7 @@ public class RespondReactorTests
         // Arrange
         var totalActionInvokes = 0;
 
-        var sut = new RespondSubscription<string>(
+        var sut = new RespondSubscription<int, string>(
             It.IsAny<Guid>(),
             It.IsAny<string>(),
             onError: e =>
@@ -172,9 +187,9 @@ public class RespondReactorTests
     }
 
     [Theory]
-    [InlineData("test-value", "4ff67e7b-bdda-4e0c-b34c-0b32270c336d", "test-value - 4ff67e7b-bdda-4e0c-b34c-0b32270c336d")]
-    [InlineData(null, "4ff67e7b-bdda-4e0c-b34c-0b32270c336d", "4ff67e7b-bdda-4e0c-b34c-0b32270c336d")]
-    [InlineData("", "4ff67e7b-bdda-4e0c-b34c-0b32270c336d", "4ff67e7b-bdda-4e0c-b34c-0b32270c336d")]
+    [InlineData("test-value", "87e99bdc-a972-427a-90be-f2c07c4f9aef", "test-value - 87e99bdc-a972-427a-90be-f2c07c4f9aef")]
+    [InlineData(null, "87e99bdc-a972-427a-90be-f2c07c4f9aef", "87e99bdc-a972-427a-90be-f2c07c4f9aef")]
+    [InlineData("", "87e99bdc-a972-427a-90be-f2c07c4f9aef", "87e99bdc-a972-427a-90be-f2c07c4f9aef")]
     public void ToString_WhenInvoked_ReturnsCorrectResult(
         string name,
         string guid,
@@ -183,9 +198,7 @@ public class RespondReactorTests
         // Arrange
         var id = new Guid(guid);
 
-        var sut = new RespondSubscription<It.IsAnyType>(
-            respondId: id,
-            name: name);
+        var sut = new RespondSubscription<It.IsAnyType, It.IsAnyType>(id, name);
 
         // Act
         var actual = sut.ToString();
